+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cpp: Controller pipeline run by the spine after observers
 - envs: Add internal `pipelines` submodule for spine pipelines
 - examples: Trying a PyBullet simulation environment
+- model: List joint names in `Model` class
 - utils: Joystick utility class
 
 ### Changed